fix(lens): the converge and its reconciliation are one verb
`convergeAndReconcile(ep, hooks)` moves into its own leaf with the verdict it
produces; `reconcileConverged` is private to it. The boot, the healer's
re-converge and the suite all take the same verb instead of re-assembling the
triple, and the claim that covers the fail-closed recycle is taken once, inside.

- hickey F3 — `converge → outcomeAdopted → reconcileConverged` was hand-assembled
  at three sites with `ep` passed twice and a derived `adopted` boolean beside it;
  reconciling endpoint A against endpoint B's outcome is no longer spellable, and
  the test's hand-rolled `convergeWithHooks` deletes.
- hickey F3 — the hooks are their own two-key type, so what the step reads is what
  its signature says rather than the whole `ensureLocalEndpoint` opts bag.
- lowy F4 — `ConvergeVerdict` was declared at a CONSUMER (`linkLoss.ts`) and
  imported backwards by its producer, and `reconcileConverged` was exported from
  the composition root "for the test". Both now live in
  `ptyHost/reconcileConverged.ts`; every arrow points down.

Raised by the lowy ∥ hickey lens review. Not pushed or merged.
